feat: use fetch and restructure lambda deps
