+++ package.json
@@ -19,10 +19,8 @@
   "dependencies": {
     "@apollo/react-hooks": "^3.1.3",
     "apollo-boost": "^0.4.4",
-    "apollo-server-lambda": "^2.9.6",
     "babel-plugin-styled-components": "^1.10.0",
     "dotenv": "^8.2.0",
-    "faunadb": "^2.9.2",
     "gatsby": "^2.10.5",
     "gatsby-image": "^2.0.29",
     "gatsby-plugin-catch-links": "^2.0.11",
