+++ src/lambda/graphql/faunaSchema.graphql
